Move postprocessing to new module.
Postprocessing operations such as audio normalization and writing tags and artwork to audio/video files have been moved to a separate postprocessing module. This required refactoring of Context into a separate module to avoid circular dependencies.
